HW23 Move a resolver
